feat: api keys for endpoints (#4)
## fixes KILTProtocol/ticket#3024

Adds API-Keys for the RPC-Endpoints from Dwellir and OnFinality. 

Makes the requests quota better, but still gets exceeded after a while. 
Not a real problem because it just changes between endpoints and gives
time for quota of the previous one to reset.


Second merge of the same brach, but this time squashed.