feat: better block ID (#46)
## fixes KILTProtocol/ticket#NoTicket

## Context
The block numbers are used as the IDs of the block _entities_ (data
structure). The IDs are always saved as character strings, what makes
"100" to be smaller than "7".

The biggest block number allowed by the blockchain has 20 digits (2^64).

## Function of this Pull Request 
Normalizes the IDs of the Blocks identities to 9 digits by prefixing
with insignificant zeros.

This would allowed to properly sort blocks entities by using their block
numbers.

It would take the [Blockchain 380,25 years to make 1 billion
blocks](https://duckduckgo.com/?q=10%5E9%2F(365%2C25*24*60*60%2F12)&ia=calculator).

For this reason, I think that 9 digits is enough for now. On 377 years
it could be expanded to more digits.

_Example_:
from "1234567" to "001234567".