feat: unit tests (#42)
## fixes KILTProtocol/ticket#3518

Takes advantage of the [SubQuery testing
framework](https://academy.subquery.network/indexer/build/testing.html#the-subquery-testing-framework)
to test the behaviour of the mapping handlers. It interacts with
specific blocks of the production KILT blockchain and validates how the
data is being indexed. Only the dependent entities (past data) are
mocked.

On this branch: 
- The github workflow for the pull requests (`pr.yml`) was modified to
successfully run the subquery tests.
- The tests for the handlers were written.
- Small bugs found during testing were fixed. 
- A _Testing_ section was added to the `Readme.md`.

---------

Co-authored-by: Andrés <105802444+kilted-andres@users.noreply.github.com>